Moved pkg/core/utils/variablesreplacer to pkg/varsub
