+++ pkg/varsub/varsub.go
@@ -1,4 +1,4 @@
-package utils
+package varsub
 
 import (
 	"regexp"
